Move form status, pass accessor to status func
