+++ src/components/Form/FormStatus/FormStatus.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { InlineIcon } from '..';
+import { InlineIcon } from '../..';
 
 export type FormStatusProps = {
   type: 'warning' | 'error' | 'success';
