+++ src/components/Form/index.ts
@@ -4,3 +4,4 @@ export * from './FormField';
 export * from './FormFieldSet';
 export * from './FormSubmitButton';
 export * from './ObjectDriller';
+export * from './FormStatus/FormStatus'
