+++ src/components/_index.scss
@@ -4,7 +4,7 @@
 @import './ButtonWithDropdown';
 @import './TextWithIcon';
 @import './FieldSet';
-@import './FormStatus';
+@import './Form/FormStatus';
 @import './Header';
 @import './InlineIcon';
 @import './Legend';
