+++ src/components/index.ts
@@ -14,7 +14,6 @@ export * from './ErrorBoundary/ErrorBoundary';
 export * from './FieldSet/FieldSet';
 export * from './FileInput/FileInput';
 export * from './Form';
-export * from './FormStatus/FormStatus';
 export * from './Header/Header';
 export * from './Header/NavLink';
 export * from './Header/NavDropdown';
